refactor: app 라우팅 레이어와 서비스/공통 로직 구조 분리
