+++ src/app/api/weather/route.ts
@@ -1,8 +1,8 @@
-import { NextResponse } from 'next/server';
+﻿import { NextResponse } from 'next/server';
 
-import { fetchAirPollution, fetchWeather } from '@/app/services/weather/openWeather';
-import { normalizeAir, normalizeWeather } from '@/app/services/weather/normalizeWeather';
-import { secondsUntilNextKstBoundary } from '@/app/services/weather/timeSlotCache';
+import { fetchAirPollution, fetchWeather } from '@/services/weather/openWeather';
+import { normalizeAir, normalizeWeather } from '@/services/weather/normalizeWeather';
+import { secondsUntilNextKstBoundary } from '@/services/weather/timeSlotCache';
 
 // 서울 기본 좌표 (위치 미사용 시 fallback)
 const SEOUL_LAT = 37.5665;
