+++ src/domain/Mypage/AccountSetting/FaqModal/FaqModal.tsx
@@ -1,11 +1,11 @@
-import { useCallback, useRef, useState } from 'react';
+﻿import { useCallback, useRef, useState } from 'react';
 import { toast } from 'react-toastify';
 
 import Modal from '@/shared/components/Modal';
 import BaseInput from '@/shared/components/Input/BaseInput';
 import Button from '@/shared/components/Button';
 
-import { faqServiceClient } from '@/app/services/backend/faq.api';
+import { faqServiceClient } from '@/services/backend/faq.api';
 import type { FaqModalProps } from './types';
 
 import styles from './FaqModal.module.scss';
