+++ src/domain/Mypage/FoodDotSelectionCard/FoodDotSelectionCard.tsx
@@ -1,4 +1,4 @@
-'use client';
+﻿'use client';
 
 import { useState } from 'react';
 import { Check, Utensils } from 'lucide-react';
@@ -8,7 +8,7 @@ import { toast } from 'react-toastify';
 
 import Modal from '@/shared/components/Modal';
 
-import { addFoodDot, getMyFoodDots, removeFoodDot } from '@/app/services/backend/users.api';
+import { addFoodDot, getMyFoodDots, removeFoodDot } from '@/services/backend/users.api';
 
 import { FOOD_DOTS } from '@/domain/Mypage/constants/foodDots';
