+++ src/services/backend/favorites.api.ts
@@ -1,5 +1,5 @@
-import { fetcherClient } from '@/app/lib/fetcher-client';
-import { fetcherServer } from '@/app/lib/fetcher-server';
+import { fetcherClient } from '@/lib/fetcher-client';
+import { fetcherServer } from '@/lib/fetcher-server';
 
 interface Fetcher {
   <T>(url: string, options?: RequestInit): Promise<T>;
@@ -7,14 +7,14 @@ interface Fetcher {
 
 class FavoritesService {
   constructor(private fetcher: Fetcher) {}
-  // 찜 목록 조회
+  // 나의 찜 목록 조회
   getMyFavorites(): Promise<Favorite.GetMyFavoritesRes[]> {
     return this.fetcher<Favorite.GetMyFavoritesRes[]>('/favorites/list', {
       method: 'GET',
     });
   }
 
-  // 찜 기반 선호도 분석 조회
+  // 나의 선호도 조회
   getMyPreference(): Promise<Favorite.PreferenceResult> {
     return this.fetcher<Favorite.PreferenceResult>('/favorites/preference', {
       method: 'GET',
