+++ src/services/weather/normalizeWeather.ts
