+++ src/services/weather/timeSlotCache.ts
